@@ -1,16 +1,15 @@
 package service
 
 import (
-	"crypto/subtle"
 	"encoding/json"
 	"encoding/xml"
 	"fmt"
-	"github.com/gorilla/mux"
+	"github.com/gin-gonic/gin"
 	"github.com/je4/FairService/v2/pkg/fair"
 	"github.com/je4/FairService/v2/pkg/model/dataciteModel"
 	"github.com/je4/FairService/v2/pkg/model/dcmi"
-	"github.com/op/go-logging"
-	"io/ioutil"
+	"github.com/je4/utils/v2/pkg/zLogger"
+	"io"
 	"net/http"
 	"regexp"
 	"strings"
@@ -23,28 +22,28 @@ type FairResultStatus struct {
 	ArchiveItems []*fair.ArchiveItem `json:"archiveitems,omitempty"`
 }
 
-func sendCreateResult(log *logging.Logger, w http.ResponseWriter, t string, message string, item *fair.ItemData) {
+func sendResult(log zLogger.ZLogger, ctx *gin.Context, status int, message string, item *fair.ItemData) {
 	if item != nil {
-		if t == "ok" {
-			log.Infof(fmt.Sprintf("%s: %s", message, item.UUID))
+		if status == http.StatusOK {
+			log.Info().Msgf("%s: %s", message, item.UUID)
 		} else {
-			log.Error(fmt.Sprintf("%s: %s", message, item.UUID))
+			log.Error().Msgf("%s: %s", message, item.UUID)
 		}
 	} else {
-		if t == "ok" {
-			log.Infof(fmt.Sprintf("%s", message))
+		if status == http.StatusOK {
+			log.Info().Msgf("%s", message)
 		} else {
-			log.Error(fmt.Sprintf("%s", message))
+			log.Error().Msgf("%s", message)
 		}
 	}
-	w.Header().Set("Content-type", "text/json")
-	data, _ := json.MarshalIndent(FairResultStatus{Status: t, Message: message, Item: item}, "", "  ")
-	w.Write(data)
+	message = fmt.Sprintf("%s: %s", ctx.HandlerName(), message)
+	ctx.JSON(status, FairResultStatus{Status: http.StatusText(status), Message: message, Item: item})
 }
 
-func BasicAuth(w http.ResponseWriter, r *http.Request, username, password, realm string) bool {
+/*
+func BasicAuth(ctx *gin.Context, username, password, realm string) bool {
 
-	user, pass, ok := r.BasicAuth()
+	user, pass, ok := ctx.Request.BasicAuth()
 
 	if !ok || subtle.ConstantTimeCompare([]byte(user), []byte(username)) != 1 || subtle.ConstantTimeCompare([]byte(pass), []byte(password)) != 1 {
 		w.Header().Set("WWW-Authenticate", `Basic realm="`+realm+`"`)
@@ -56,72 +55,110 @@ func BasicAuth(w http.ResponseWriter, r *http.Request, username, password, realm
 	return true
 }
 
-func (s *Server) redirectHandler(w http.ResponseWriter, req *http.Request) {
-	vars := mux.Vars(req)
-	pName := vars["partition"]
-	uuidStr := vars["uuid"]
+*/
 
-	part, err := s.fair.GetPartition(pName)
-	if err != nil {
-		w.WriteHeader(http.StatusNotFound)
-		w.Header().Set("Content-type", "text/plain")
-		w.Write([]byte(fmt.Sprintf("partition [%s] not found", pName)))
-		return
+func (s *Server) resolverHandler(ctx *gin.Context) {
+	pid := strings.Trim(ctx.Param("pid"), "/")
+	if strings.HasPrefix(strings.ToLower(pid), "ark:") {
+		uuid, components, variants, err := s.fair.ARKResolveUUID(pid)
+		if err != nil {
+			sendResult(s.log, ctx, http.StatusNotFound, fmt.Sprintf("cannot resolve %s: %v", pid, err), nil)
+			return
+		}
+		ctx.Set("uuid", uuid)
+		suffix := components
+		if variants != "" {
+			if suffix != "" {
+				suffix += "." + variants
+			} else {
+				suffix = variants
+			}
+		}
+		ctx.Set("suffix", suffix)
+		s.redirectHandler(ctx)
+	}
+}
+
+func (s *Server) redirectHandler(ctx *gin.Context) {
+	pName := ctx.Param("partition") + ctx.GetString("partition")
+	uuidStr := ctx.Param("uuid") + ctx.GetString("uuid")
+	suffix := ctx.Param("suffix") + ctx.GetString("suffix")
+
+	var part *fair.Partition
+	var err error
+	if pName != "" {
+		part, err = s.fair.GetPartition(pName)
+		if err != nil {
+			ctx.AbortWithError(http.StatusNotFound, fmt.Errorf("%s: partition [%s] not found", ctx.HandlerName(), pName))
+			return
+		}
 	}
 	data, err := s.fair.GetItem(part, uuidStr)
 	if err != nil {
-		w.WriteHeader(http.StatusNotFound)
-		w.Header().Set("Content-type", "text/plain")
-		w.Write([]byte(fmt.Sprintf("error loading item %s/%s: %v", pName, uuidStr, err)))
+		ctx.AbortWithError(http.StatusNotFound, fmt.Errorf("error loading item %s/%s: %v", pName, uuidStr, err))
+		return
+	}
+	if part == nil {
+		part, err = s.fair.GetPartition(data.Partition)
+		if err != nil {
+			ctx.AbortWithError(http.StatusNotFound, fmt.Errorf("%s: partition [%s] not found", ctx.HandlerName(), data.Partition))
+			return
+		}
+	}
+	if data.Status == fair.DataStatusDeletedMeta {
+		tpl := s.templates["detail"]
+		if err := tpl.Execute(ctx.Writer, struct {
+			BaseURL string
+			Part    *fair.Partition
+			Data    *fair.ItemData
+		}{BaseURL: part.AddrExt + "/", Part: part, Data: data}); err != nil {
+			ctx.AbortWithError(http.StatusInternalServerError, fmt.Errorf("error executing template %s in partition %s: %v", "partition", pName, err))
+			return
+		}
 		return
 	}
 	if data.Status != fair.DataStatusActive {
 		tpl := s.templates["deleted"]
-		if err := tpl.Execute(w, struct {
-			Part *fair.Partition
-			Data *fair.ItemData
-		}{Part: part, Data: data}); err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			w.Header().Set("Content-type", "text/plain")
-			w.Write([]byte(fmt.Sprintf("error executing template %s in partition %s: %v", "partition", pName, err)))
+		if err := tpl.Execute(ctx.Writer, struct {
+			BaseURL string
+			Part    *fair.Partition
+			Data    *fair.ItemData
+		}{BaseURL: part.AddrExt + "/", Part: part, Data: data}); err != nil {
+			ctx.AbortWithError(http.StatusInternalServerError, fmt.Errorf("error executing template %s in partition %s: %v", "partition", pName, err))
 			return
 		}
-
-	} else {
-		source, err := s.fair.GetSourceByName(part, data.Source)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			w.Header().Set("Content-type", "text/plain")
-			w.Write([]byte(fmt.Sprintf("error loading source %s for item %s/%s: %v", data.Source, pName, uuidStr, err)))
-			return
-		}
-
-		targetURL := strings.Replace(source.DetailURL, "{signature}", data.Signature, -1)
-
-		http.Redirect(w, req, targetURL, 301)
+		return
 	}
-}
-
-func (s *Server) detailHandler(w http.ResponseWriter, req *http.Request) {
-	if !BasicAuth(w, req, s.name, s.password, "FAIR Service") {
+	source, err := s.fair.GetSourceByName(data.Partition, data.Source)
+	if err != nil {
+		ctx.AbortWithError(http.StatusInternalServerError, fmt.Errorf("error loading source %s for item %s/%s: %v", data.Source, pName, uuidStr, err))
 		return
 	}
 
-	vars := mux.Vars(req)
-	pName := vars["partition"]
-	uuidStr := vars["uuid"]
+	var targetURL string
+	if data.URL != "" {
+		targetURL = data.URL
+	} else {
+		targetURL = strings.Replace(source.DetailURL, "{signature}", data.Signature, -1)
+	}
+	targetURL += suffix
+	ctx.Redirect(http.StatusMovedPermanently, targetURL)
+}
+
+func (s *Server) detailHandler(ctx *gin.Context) {
+	pName := ctx.Param("partition")
+	uuidStr := ctx.Param("uuid")
 
 	part, err := s.fair.GetPartition(pName)
 	if err != nil {
-		w.WriteHeader(http.StatusNotFound)
-		w.Header().Set("Content-type", "text/plain")
-		w.Write([]byte(fmt.Sprintf("partition [%s] not found", pName)))
+		ctx.AbortWithError(http.StatusNotFound, fmt.Errorf("%s: partition [%s] not found", ctx.HandlerName(), pName))
 		return
 	}
 
 	doiError := ""
 	message := ""
-	if _, ok := req.URL.Query()["createdoi"]; ok {
+
+	if _, ok := ctx.GetQuery("createdoi"); ok {
 		targetUrl := fmt.Sprintf("%s/redir/%s", part.AddrExt, uuidStr)
 		_, err := s.fair.CreateDOI(part, uuidStr, targetUrl)
 		if err != nil {
@@ -133,298 +170,248 @@ func (s *Server) detailHandler(w http.ResponseWriter, req *http.Request) {
 
 	data, err := s.fair.GetItem(part, uuidStr)
 	if err != nil {
-		w.WriteHeader(http.StatusNotFound)
-		w.Header().Set("Content-type", "text/plain")
-		w.Write([]byte(fmt.Sprintf("error loading item %s/%s: %v", pName, uuidStr, err)))
+		ctx.AbortWithError(http.StatusNotFound, fmt.Errorf("error loading item %s/%s: %v", pName, uuidStr, err))
 		return
 	}
 	tpl := s.templates["detail"]
-	if err := tpl.Execute(w, struct {
+	if err := tpl.Execute(ctx.Writer, struct {
+		BaseURL string
 		Error   string
 		Message string
 		Part    *fair.Partition
 		Data    *fair.ItemData
-	}{Error: doiError, Message: message, Part: part, Data: data}); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Header().Set("Content-type", "text/plain")
-		w.Write([]byte(fmt.Sprintf("error executing template %s in partition %s: %v", "partition", pName, err)))
+	}{BaseURL: part.AddrExt + "/", Error: doiError, Message: message, Part: part, Data: data}); err != nil {
+		ctx.AbortWithError(http.StatusInternalServerError, fmt.Errorf("error executing template %s in partition %s: %v", "partition", pName, err))
 		return
 	}
 }
 
-func (s *Server) partitionHandler(w http.ResponseWriter, req *http.Request) {
-	vars := mux.Vars(req)
-	pName := vars["partition"]
+func (s *Server) partitionHandler(ctx *gin.Context) {
+	pName := ctx.Param("partition")
 
 	part, err := s.fair.GetPartition(pName)
 	if err != nil {
-		w.WriteHeader(http.StatusNotFound)
-		w.Header().Set("Content-type", "text/plain")
-		w.Write([]byte(fmt.Sprintf("partition [%s] not found", pName)))
+		ctx.AbortWithError(http.StatusNotFound, fmt.Errorf("%s: partition [%s] not found", ctx.HandlerName(), pName))
 		return
 	}
 
 	tpl := s.templates["partition"]
-	if err := tpl.Execute(w, part); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Header().Set("Content-type", "text/plain")
-		w.Write([]byte(fmt.Sprintf("error executing template %s in partition %s: %v", "partition", pName, err)))
+	if err := tpl.Execute(ctx.Writer, struct {
+		BaseURL string
+		Part    *fair.Partition
+	}{
+		BaseURL: part.AddrExt + "/",
+		Part:    part,
+	}); err != nil {
+		ctx.AbortWithError(http.StatusInternalServerError, fmt.Errorf("error executing template %s in partition %s: %v", "partition", pName, err))
 		return
 	}
 }
 
-func (s *Server) partitionOAIHandler(w http.ResponseWriter, req *http.Request) {
-	vars := mux.Vars(req)
-	pName := vars["partition"]
+func (s *Server) partitionOAIHandler(ctx *gin.Context) {
+	pName := ctx.Param("partition")
 
 	part, err := s.fair.GetPartition(pName)
 	if err != nil {
-		w.WriteHeader(http.StatusNotFound)
-		w.Header().Set("Content-type", "text/plain")
-		w.Write([]byte(fmt.Sprintf("partition [%s] not found", pName)))
+		ctx.AbortWithError(http.StatusNotFound, fmt.Errorf("%s: partition [%s] not found", ctx.HandlerName(), pName))
 		return
 	}
 
 	tpl := s.templates["oai"]
-	if err := tpl.Execute(w, part); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Header().Set("Content-type", "text/plain")
-		w.Write([]byte(fmt.Sprintf("error executing template %s in partition %s: %v", "partition", pName, err)))
+	if err := tpl.Execute(ctx.Writer, struct {
+		BaseURL string
+		Part    *fair.Partition
+	}{BaseURL: part.AddrExt + "/", Part: part}); err != nil {
+		ctx.AbortWithError(http.StatusInternalServerError, fmt.Errorf("error executing template %s in partition %s: %v", "partition", pName, err))
 		return
 	}
 }
 
-func (s *Server) itemHandler(w http.ResponseWriter, req *http.Request) {
-
-	vars := mux.Vars(req)
-	pName := vars["partition"]
-	uuidStr := vars["uuid"]
-	outputType := vars["outputType"]
+func (s *Server) itemHandler(ctx *gin.Context) {
+	pName := ctx.Param("partition")
+	uuidStr := ctx.Param("uuid")
+	outputType := ctx.Param("outputType")
 	if outputType == "" {
 		outputType = "json"
 	}
 
 	part, err := s.fair.GetPartition(pName)
 	if err != nil {
-		w.WriteHeader(http.StatusNotFound)
-		w.Header().Set("Content-type", "text/plain")
-		w.Write([]byte(fmt.Sprintf("partition [%s] not found", pName)))
+		ctx.AbortWithError(http.StatusNotFound, fmt.Errorf("%s: partition [%s] not found", ctx.HandlerName(), pName))
 		return
 	}
 
 	data, err := s.fair.GetItem(part, uuidStr)
 	if err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("error loading item %v: %v", uuidStr, err), nil)
+		sendResult(s.log, ctx, http.StatusInternalServerError, fmt.Sprintf("error loading item %v: %v", uuidStr, err), nil)
 		return
 	}
 	if data == nil {
-		w.WriteHeader(http.StatusNotFound)
-		w.Header().Set("Content-type", "text/plain")
-		w.Write([]byte(fmt.Sprintf("item [%s] not found in partition %s", uuidStr, pName)))
+		ctx.AbortWithError(http.StatusNotFound, fmt.Errorf("item [%s] not found in partition %s", uuidStr, pName))
 		return
 	}
 	if data.Access != fair.DataAccessOpenAccess && data.Access != fair.DataAccessPublic {
-		w.WriteHeader(http.StatusForbidden)
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("no public access for %v: %v", uuidStr, data.Access), nil)
+		sendResult(s.log, ctx, http.StatusForbidden, fmt.Sprintf("no public access for %v: %v", uuidStr, data.Access), nil)
 		return
 	}
 	if data.Status != fair.DataStatusActive {
-		w.WriteHeader(http.StatusForbidden)
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("status of %v not active: %v", uuidStr, data.Status), nil)
+		ctx.AbortWithError(http.StatusForbidden, fmt.Errorf("status of %v not active: %v", uuidStr, data.Status))
 		return
 	}
 	switch outputType {
 	case "json":
-		w.Header().Set("Content-type", "text/json")
-		enc := json.NewEncoder(w)
-		enc.SetIndent("", "  ")
-		if err := enc.Encode(data); err != nil {
-			sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot marshal data for %v", uuidStr), nil)
-			return
-		}
+		ctx.JSON(http.StatusOK, data)
 		return
 	case "dcmi":
 		dcmiData := &dcmi.DCMI{}
 		dcmiData.InitNamespace()
 		dcmiData.FromCore(data.Metadata)
-		w.Header().Set("Content-type", "text/xml")
-		enc := xml.NewEncoder(w)
-		enc.Indent("", "  ")
-		if err := enc.Encode(dcmiData); err != nil {
-			sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot marshal data for %v", uuidStr), nil)
-			return
-		}
+		ctx.XML(http.StatusOK, dcmiData)
 		return
 	case "datacite":
 		dataciteData := &dataciteModel.DataCite{}
 		dataciteData.InitNamespace()
 		dataciteData.FromCore(data.Metadata)
-		w.Header().Set("Content-type", "text/xml")
-		enc := xml.NewEncoder(w)
-		enc.Indent("", "  ")
-		if err := enc.Encode(dataciteData); err != nil {
-			sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot marshal data for %v", uuidStr), nil)
-			return
-		}
+		ctx.XML(http.StatusOK, dataciteData)
 		return
 	default:
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("invalid output type %s for %v", outputType, uuidStr), nil)
+		sendResult(s.log, ctx, http.StatusBadRequest, fmt.Sprintf("invalid output type %s for %v", outputType, uuidStr), nil)
 		return
 	}
 }
 
-func (s *Server) createDOIHandler(w http.ResponseWriter, req *http.Request) {
-	if !BasicAuth(w, req, s.name, s.password, "FAIR Service") {
-		return
-	}
+func (s *Server) createDOIHandler(ctx *gin.Context) {
 
-	vars := mux.Vars(req)
-	pName := vars["partition"]
-	uuidStr := vars["uuid"]
+	pName := ctx.Param("partition")
+	uuidStr := ctx.Param("uuid")
 
 	part, err := s.fair.GetPartition(pName)
 	if err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot get partition %s for %v", pName, uuidStr), nil)
+		sendResult(s.log, ctx, http.StatusNotFound, fmt.Sprintf("cannot get partition %s for %v", pName, uuidStr), nil)
 		return
 	}
 
 	targetUrl := fmt.Sprintf("%s/redir/%s", part.AddrExt, uuidStr)
 	doiResult, err := s.fair.CreateDOI(part, uuidStr, targetUrl)
 	if err != nil {
-		sendCreateResult(s.log, w, "error", err.Error(), nil)
+		sendResult(s.log, ctx, http.StatusInternalServerError, err.Error(), nil)
 		return
 	}
 
-	w.Header().Set("Content-type", "text/json")
-	enc := json.NewEncoder(w)
-	enc.SetIndent("", "  ")
-	if err := enc.Encode(doiResult); err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot marshal data for %v", uuidStr), nil)
-		return
-	}
+	ctx.JSON(http.StatusOK, doiResult)
 	return
 }
 
-func (s *Server) startUpdateHandler(w http.ResponseWriter, req *http.Request) {
+func (s *Server) startUpdateHandler(ctx *gin.Context) {
 
-	vars := mux.Vars(req)
-	pName := vars["partition"]
+	pName := ctx.Param("partition")
 
 	part, err := s.fair.GetPartition(pName)
 	if err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot get partition %s", pName), nil)
+		sendResult(s.log, ctx, http.StatusNotFound, fmt.Sprintf("cannot get partition %s", pName), nil)
 		return
 	}
 
-	decoder := json.NewDecoder(req.Body)
 	var data fair.SourceData
-	if err := decoder.Decode(&data); err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot parse request body: %v", err), nil)
+	if err := ctx.ShouldBindJSON(&data); err != nil {
+		sendResult(s.log, ctx, http.StatusBadRequest, fmt.Sprintf("cannot parse request body: %v", err), nil)
 		return
 	}
 
 	if err := s.fair.StartUpdate(part, data.Source); err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot start update for %s on %s: %v", data.Source, pName, err), nil)
+		sendResult(s.log, ctx, http.StatusInternalServerError, fmt.Sprintf("cannot start update for %s on %s: %v", data.Source, pName, err), nil)
 		return
 	}
-	sendCreateResult(s.log, w, "ok", fmt.Sprintf("starting update for %s on %s", data.Source, pName), nil)
+	sendResult(s.log, ctx, http.StatusOK, fmt.Sprintf("starting update for %s on %s", data.Source, pName), nil)
 }
 
-func (s *Server) pingHandler(w http.ResponseWriter, req *http.Request) {
-	sendCreateResult(s.log, w, "ok", "pong", nil)
+func (s *Server) pingHandler(ctx *gin.Context) {
+	sendResult(s.log, ctx, http.StatusOK, "pong", nil)
 }
 
-func (s *Server) endUpdateHandler(w http.ResponseWriter, req *http.Request) {
-
-	vars := mux.Vars(req)
-	pName := vars["partition"]
+func (s *Server) endUpdateHandler(ctx *gin.Context) {
+	pName := ctx.Param("partition")
 
 	part, err := s.fair.GetPartition(pName)
 	if err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot get partition %s", pName), nil)
+		sendResult(s.log, ctx, http.StatusNotFound, fmt.Sprintf("cannot get partition %s", pName), nil)
 		return
 	}
 
-	decoder := json.NewDecoder(req.Body)
 	var data fair.SourceData
-	if err := decoder.Decode(&data); err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot parse request body: %v", err), nil)
+	if err := ctx.ShouldBindJSON(&data); err != nil {
+		sendResult(s.log, ctx, http.StatusBadRequest, fmt.Sprintf("cannot parse request body: %v", err), nil)
 		return
 	}
 
 	if err := s.fair.EndUpdate(part, data.Source); err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot end update for %s on %s: %v", data.Source, pName, err), nil)
+		sendResult(s.log, ctx, http.StatusInternalServerError, fmt.Sprintf("cannot end update for %s on %s: %v", data.Source, pName, err), nil)
 		return
 	}
-	sendCreateResult(s.log, w, "ok", fmt.Sprintf("end update for %s on %s", data.Source, pName), nil)
+	sendResult(s.log, ctx, http.StatusOK, fmt.Sprintf("end update for %s on %s", data.Source, pName), nil)
 }
 
-func (s *Server) abortUpdateHandler(w http.ResponseWriter, req *http.Request) {
-	vars := mux.Vars(req)
-	pName := vars["partition"]
+func (s *Server) abortUpdateHandler(ctx *gin.Context) {
+	pName := ctx.Param("partition")
 
 	part, err := s.fair.GetPartition(pName)
 	if err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot get partition %s", pName), nil)
+		sendResult(s.log, ctx, http.StatusNotFound, fmt.Sprintf("cannot get partition %s", pName), nil)
 		return
 	}
 
-	decoder := json.NewDecoder(req.Body)
 	var data fair.SourceData
-	if err := decoder.Decode(&data); err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot parse request body: %v", err), nil)
+	if err := ctx.ShouldBindJSON(&data); err != nil {
+		sendResult(s.log, ctx, http.StatusBadRequest, fmt.Sprintf("cannot parse request body: %v", err), nil)
 		return
 	}
 
 	if err := s.fair.AbortUpdate(part, data.Source); err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot end update for %s on %s: %v", data.Source, pName, err), nil)
+		sendResult(s.log, ctx, http.StatusInternalServerError, fmt.Sprintf("cannot end update for %s on %s: %v", data.Source, pName, err), nil)
 		return
 	}
-	sendCreateResult(s.log, w, "ok", fmt.Sprintf("abort update for %s on %s", data.Source, pName), nil)
+	sendResult(s.log, ctx, http.StatusOK, fmt.Sprintf("abort update for %s on %s", data.Source, pName), nil)
 }
 
-func (s *Server) originalDataReadHandler(w http.ResponseWriter, req *http.Request) {
-	vars := mux.Vars(req)
-	pName := vars["partition"]
-	uuidStr := vars["uuid"]
+func (s *Server) originalDataReadHandler(ctx *gin.Context) {
+	pName := ctx.Param("partition")
+	uuidStr := ctx.Param("uuid")
 
 	part, err := s.fair.GetPartition(pName)
 	if err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot get partition %s", pName), nil)
+		sendResult(s.log, ctx, http.StatusNotFound, fmt.Sprintf("cannot get partition %s", pName), nil)
 		return
 	}
 
 	data, t, err := s.fair.GetOriginalData(part, uuidStr)
 	if err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot read original data: %v", err), nil)
+		sendResult(s.log, ctx, http.StatusInternalServerError, fmt.Sprintf("cannot read original data: %v", err), nil)
 		return
 	}
 	switch t {
 	case "XML":
-		w.Header().Set("Content-Type", "text/xml")
+		ctx.Header("Content-Type", "text/xml")
 	case "JSON":
-		w.Header().Set("Content-Type", "application/json")
+		ctx.Header("Content-Type", "application/json")
 	default:
-		w.Header().Set("Content-Type", "text/plain")
+		ctx.Header("Content-Type", "text/plain")
 	}
-	w.Write(data)
+	ctx.Writer.Write(data)
 }
 
-func (s *Server) originalDataWriteHandler(w http.ResponseWriter, req *http.Request) {
-	vars := mux.Vars(req)
-	pName := vars["partition"]
-	uuidStr := vars["uuid"]
+func (s *Server) originalDataWriteHandler(ctx *gin.Context) {
+	pName := ctx.Param("partition")
+	uuidStr := ctx.Param("uuid")
 
 	part, err := s.fair.GetPartition(pName)
 	if err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot get partition %s", pName), nil)
+		sendResult(s.log, ctx, http.StatusNotFound, fmt.Sprintf("cannot get partition %s", pName), nil)
 		return
 	}
 
-	bdata, err := ioutil.ReadAll(req.Body)
+	bdata, err := io.ReadAll(ctx.Request.Body)
 	if err != nil {
-		s.log.Errorf("cannot read request body: %v", err)
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot read request body: %v", err), nil)
+		sendResult(s.log, ctx, http.StatusInternalServerError, fmt.Sprintf("cannot read request body: %v", err), nil)
 		return
 	}
 
@@ -441,82 +428,55 @@ func (s *Server) originalDataWriteHandler(w http.ResponseWriter, req *http.Reque
 	}
 
 	if err := s.fair.SetOriginalData(part, uuidStr, t, bdata); err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot set original data for %s: %v", uuidStr, err), nil)
+		sendResult(s.log, ctx, http.StatusInternalServerError, fmt.Sprintf("cannot set original data for %s: %v", uuidStr, err), nil)
 		return
 	}
-	sendCreateResult(s.log, w, "ok", fmt.Sprintf("original metadata for %s stored", uuidStr), nil)
+	sendResult(s.log, ctx, http.StatusOK, fmt.Sprintf("original data for %s stored", uuidStr), nil)
 }
 
-func (s *Server) createHandler(w http.ResponseWriter, req *http.Request) {
-	vars := mux.Vars(req)
-	pName := vars["partition"]
+func (s *Server) createHandler(ctx *gin.Context) {
+	pName := ctx.Param("partition")
 
 	part, err := s.fair.GetPartition(pName)
 	if err != nil {
-		w.WriteHeader(http.StatusNotFound)
-		w.Header().Set("Content-type", "text/plain")
-		w.Write([]byte(fmt.Sprintf("partition [%s] not found", pName)))
+		sendResult(s.log, ctx, http.StatusNotFound, fmt.Sprintf("partition [%s] not found", pName), nil)
 		return
 	}
 
 	var data = &fair.ItemData{}
 
-	/*
-		decoder := json.NewDecoder(req.Body)
-		err := decoder.Decode(&data)
-	*/
-	bdata, err := ioutil.ReadAll(req.Body)
-	if err != nil {
-		s.log.Errorf("cannot read request body: %v", err)
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot read request body: %v", err), nil)
-		return
-	}
-
-	if err := json.Unmarshal(bdata, data); err != nil {
-		s.log.Errorf("cannot unmarshal request body [%s]: %v", string(bdata), err)
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot unmarshal request body [%s]: %v", string(bdata), err), nil)
+	if err := ctx.ShouldBindJSON(data); err != nil {
+		sendResult(s.log, ctx, http.StatusBadRequest, fmt.Sprintf("cannot parse request body: %v", err), nil)
 		return
 	}
 
 	item, err := s.fair.CreateItem(part, data)
 	if err != nil {
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot create item: %v", err), nil)
+		sendResult(s.log, ctx, http.StatusInternalServerError, fmt.Sprintf("cannot create item: %v", err), nil)
 		return
 	}
-	sendCreateResult(s.log, w, "ok", "update done", item)
+	sendResult(s.log, ctx, http.StatusOK, "update done", item)
 	return
 }
 
-func (s *Server) setSourceHandler(w http.ResponseWriter, req *http.Request) {
-	vars := mux.Vars(req)
-	pName := vars["partition"]
+func (s *Server) setSourceHandler(ctx *gin.Context) {
+	pName := ctx.Param("partition")
 
 	var data = &fair.Source{}
-	bdata, err := ioutil.ReadAll(req.Body)
-	if err != nil {
-		s.log.Errorf("cannot read request body: %v", err)
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot read request body: %v", err), nil)
-		return
-	}
-
-	if err := json.Unmarshal(bdata, data); err != nil {
-		s.log.Errorf("cannot unmarshal request body [%s]: %v", string(bdata), err)
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot unmarshal request body [%s]: %v", string(bdata), err), nil)
+	if err := ctx.ShouldBindJSON(data); err != nil {
+		sendResult(s.log, ctx, http.StatusBadRequest, fmt.Sprintf("cannot parse request body: %v", err), nil)
 		return
 	}
 	if data.Partition != pName {
-		s.log.Errorf("source and partition do not match %s != %s", data.Partition, pName)
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("source and partition do not match %s != %s", data.Partition, pName), nil)
+		sendResult(s.log, ctx, http.StatusBadRequest, fmt.Sprintf("source and partition do not match %s != %s", data.Partition, pName), nil)
 		return
 	}
 
 	if err := s.fair.SetSource(data); err != nil {
-		s.log.Errorf("cannot create source %v: %v", data, err)
-		sendCreateResult(s.log, w, "error", fmt.Sprintf("cannot create source %v: %v", data, err), nil)
+		sendResult(s.log, ctx, http.StatusInternalServerError, fmt.Sprintf("cannot set source %v: %v", data, err), nil)
 		return
 	}
-	sendCreateResult(s.log, w, "ok", fmt.Sprintf("cannot create source %v: %v", data, err), nil)
-
+	sendResult(s.log, ctx, http.StatusOK, fmt.Sprintf("source %v set", data), nil)
 }
 
 type DataTableResult struct {
