@@ -1,14 +1,16 @@
 package fair
 
 import (
+	"context"
 	"database/sql"
 	"fmt"
 	"github.com/pkg/errors"
 )
 
 func (f *Fair) LoadSources() error {
-	sqlstr := fmt.Sprintf("SELECT sourceid, name, detailurl, description, oai_domain, partition FROM %s.source", f.dbSchema)
-	rows, err := f.db.Query(sqlstr)
+	//sqlstr := fmt.Sprintf("SELECT sourceid, name, detailurl, description, oai_domain, partition FROM %s.source", f.dbSchema)
+	sqlstr := "SELECT sourceid, name, detailurl, description, oai_domain, partition FROM source"
+	rows, err := f.db.Query(context.Background(), sqlstr)
 	if err != nil {
 		return errors.Wrapf(err, "cannot execute %s", sqlstr)
 	}
@@ -37,37 +39,42 @@ func (f *Fair) GetSourceById(partition *Partition, id int64) (*Source, error) {
 	return nil, errors.New(fmt.Sprintf("source #%v for partition %s not found", id, partition.Name))
 }
 
-func (f *Fair) GetSourceByName(partition *Partition, name string) (*Source, error) {
+func (f *Fair) GetSourceByName(pName string, name string) (*Source, error) {
 	f.sourcesMutex.RLock()
 	defer f.sourcesMutex.RUnlock()
 	for _, src := range f.sources {
-		if src.Name == name && src.Partition == partition.Name {
+		if src.Name == name && src.Partition == pName {
 			return src, nil
 		}
 	}
-	return nil, errors.New(fmt.Sprintf("source %s for partition %s not found", name, partition.Name))
+	return nil, errors.New(fmt.Sprintf("source %s for partition %s not found", name, pName))
 }
 
 func (f *Fair) SetSource(src *Source) error {
-	sqlstr := fmt.Sprintf("SELECT sourceid FROM %s.source WHERE name=$1", f.dbSchema)
+	//sqlstr := fmt.Sprintf("SELECT sourceid FROM %s.source WHERE name=$1", f.dbSchema)
+	sqlstr := "SELECT sourceid FROM source WHERE name=$1"
 	var sourceId int64
-	if err := f.db.QueryRow(sqlstr, src.Name).Scan(&sourceId); err != nil {
+	if err := f.db.QueryRow(context.Background(), sqlstr, src.Name).Scan(&sourceId); err != nil {
 		if err != sql.ErrNoRows {
 			return errors.Wrapf(err, "cannot query database - %s [%v]", sqlstr, src.ID)
 		}
 	}
 	if sourceId > 0 {
-		sqlstr = fmt.Sprintf("UPDATE %s.source "+
-			"SET name=$1, detailurl=$2, description=$3, oai_domain=$4, partition=$5 WHERE sourceid=$6 ", f.dbSchema)
+		/*		sqlstr = fmt.Sprintf("UPDATE %s.source "+
+				"SET name=$1, detailurl=$2, description=$3, oai_domain=$4, partition=$5 WHERE sourceid=$6 ", f.dbSchema)
+		*/
+		sqlstr = "UPDATE source SET name=$1, detailurl=$2, description=$3, oai_domain=$4, partition=$5 WHERE sourceid=$6 "
 		values := []interface{}{src.Name, src.DetailURL, src.Description, src.OAIDomain, src.Partition, sourceId}
-		if _, err := f.db.Exec(sqlstr, values...); err != nil {
+		if _, err := f.db.Exec(context.Background(), sqlstr, values...); err != nil {
 			return errors.Wrapf(err, "cannot insert into source database - %s [%v]", sqlstr, values)
 		}
 	} else {
-		sqlstr = fmt.Sprintf("INSERT INTO %s.source (name, detailurl, description, oai_domain, partition) "+
-			"VALUES($1, $2, $3, $4, $5)", f.dbSchema)
+		/*		sqlstr = fmt.Sprintf("INSERT INTO %s.source (name, detailurl, description, oai_domain, partition) "+
+				"VALUES($1, $2, $3, $4, $5)", f.dbSchema)
+		*/
+		sqlstr = "INSERT INTO source (name, detailurl, description, oai_domain, partition) VALUES($1, $2, $3, $4, $5)"
 		values := []interface{}{src.Name, src.DetailURL, src.Description, src.OAIDomain, src.Partition}
-		if _, err := f.db.Exec(sqlstr, values...); err != nil {
+		if _, err := f.db.Exec(context.Background(), sqlstr, values...); err != nil {
 			return errors.Wrapf(err, "cannot update source database - %s [%v]", sqlstr, values)
 		}
 	}
