@@ -112,24 +112,24 @@ const (
 )
 
 var RelatedIdentifierTypeReverse = map[string]RelatedIdentifierType{
-	"ARK":     RelatedIdentifierTypeARK,
-	"arXiv":   RelatedIdentifierTypeArXiv,
+	"ark":     RelatedIdentifierTypeARK,
+	"arxiv":   RelatedIdentifierTypeArXiv,
 	"bibcode": RelatedIdentifierTypeBibcode,
-	"DOI":     RelatedIdentifierTypeDOI,
-	"EAN13":   RelatedIdentifierTypeEAN13,
-	"EISSN":   RelatedIdentifierTypeEISSN,
-	"Handle":  RelatedIdentifierTypeHandle,
-	"IGSN":    RelatedIdentifierTypeIGSN,
-	"ISBN":    RelatedIdentifierTypeISBN,
-	"ISSN":    RelatedIdentifierTypeISSN,
-	"ISTC":    RelatedIdentifierTypeISTC,
-	"LISSN":   RelatedIdentifierTypeLISSN,
-	"LSID":    RelatedIdentifierTypeLSID,
-	"PMID":    RelatedIdentifierTypePMID,
-	"PURL":    RelatedIdentifierTypePURL,
-	"UPC":     RelatedIdentifierTypeUPC,
-	"URL":     RelatedIdentifierTypeURL,
-	"URN":     RelatedIdentifierTypeURN,
+	"doi":     RelatedIdentifierTypeDOI,
+	"ean13":   RelatedIdentifierTypeEAN13,
+	"eissn":   RelatedIdentifierTypeEISSN,
+	"handle":  RelatedIdentifierTypeHandle,
+	"igsn":    RelatedIdentifierTypeIGSN,
+	"isbn":    RelatedIdentifierTypeISBN,
+	"issn":    RelatedIdentifierTypeISSN,
+	"istc":    RelatedIdentifierTypeISTC,
+	"lissn":   RelatedIdentifierTypeLISSN,
+	"lsid":    RelatedIdentifierTypeLSID,
+	"pmid":    RelatedIdentifierTypePMID,
+	"purl":    RelatedIdentifierTypePURL,
+	"upc":     RelatedIdentifierTypeUPC,
+	"url":     RelatedIdentifierTypeURL,
+	"urn":     RelatedIdentifierTypeURN,
 	"w3id":    RelatedIdentifierTypeW3id,
 	"zotero":  RelatedIdentifierTypeZotero,
 }
